requirements.txt
snowflake-snowpark-python[pandas]
python-dotenv
snowflake-connector-python[secure-local-storage]